Add files via upload
Figure 2: Population of Turkmenistan
I built the line graph because I could visualize the data as it changed over time. The "Sum of Population" field was the next thing I added to the "Columns" shelf since it was a numerical measurement that allowed me to display the population size with time. I dragged the "Country" field into the "Filters" shelf and chose "Turkmenistan" to filter the data to only show information about Turkmenistan. I was able to narrow my attention to the particular nation of interest and get rid of any irrelevant information that would have complicated my study.